Upgrade golang.org/x/net to version 0.33.0
This was needed to fix a security vulnerability in the golang.org/x/net
package. The vulnerability was fixed to resolve the following:

An attacker can craft an input to the Parse functions that would be
processed non-linearly with respect to its length, resulting in
extremely slow parsing. This could cause a denial of service.