Clean up the look by adding better icons
These are currently ASCII icons that look good with the ligatures I am
using in my terminal. I would like to add in a new feature so that users
are able to customize the icons that are displayed in the terminal.

The extra slash in the progress bar is to make it line up better with
the playback control icons.