Make the oneline output more dynamic by updating
This makes the oneline output more dynamic by updating the song info
inline. This is done by using a carriage return character (\r) to return
the cursor to the start of the line and then printing the new song info.

This way a user is able to get the current song info without having to
run the command again.