Add scroll option to one-line output
This allows for the scrolling text to activate once our output string
has extended a certain length. This will be mostly beneficial when using
the one-line output in the TMUX status bar but also looks quite nice
when using this on the command line.

I have removed the ANSI escape code as the string `[K` gets outputted in
addition to the song information. I am assuming this is because TMUX is
already clearing the line and doing it twice leaves over that string.