Allow for player to refresh token periodically
This makes it so that the player will refresh the token every 10
minutes. This is to prevent the token from expiring while the player is
running, which would cause the player to stop working. This is a simple
way to keep the player running for long periods of time without having
to manually refresh the token.