Refactor the authentication process and add Viper
This is a big commit that refactors the authentication process that was
being done and separates it into a separate file to better manage
creating a client from the same token stored in viper for other
commands.

I still need to implement a `goroutine` that will manage the refresh
token and add some code to run before running any of my player sub
commands.