Upgrade old loading of token and client
This makes it so we use the new viper functions we have added and have
spotgo reference the values that are in the config file as a main
source. This makes it so we can have a PreRun function to get a valid
client and token before making any calls to Spotify.